feat: cleanup (replaces old component with new)
